+++ src/polymer/create-element-extension.ts
@@ -14,11 +14,11 @@
 
 import babelTraverse from 'babel-traverse';
 import * as babel from 'babel-types';
-import { Document, Severity, Warning } from 'polymer-analyzer';
+import {Document, Severity, Warning} from 'polymer-analyzer';
 
 import {registry} from '../registry';
 import {Rule} from '../rule';
-import {getDocumentContaining, stripIndentation, stripWhitespace} from '../util';
+import {getDocumentContaining, stripIndentation} from '../util';
 
 
 class CreateElementExtension extends Rule {
@@ -30,7 +30,7 @@ class CreateElementExtension extends Rule {
   async check(document: Document) {
     const warnings: Warning[] = [];
 
-    const docs = document.getFeatures({ kind: 'js-document' });
+    const docs = document.getFeatures({kind: 'js-document'});
 
     for (const doc of docs) {
       babelTraverse(doc.parsedDocument.ast, {
@@ -47,7 +47,8 @@ class CreateElementExtension extends Rule {
             message = 'Element extension via the is property is not widely supported, and is not recommended.';
           }
 
-          const containingDoc = getDocumentContaining(doc.sourceRange, document);
+          const containingDoc =
+              getDocumentContaining(doc.sourceRange, document);
           if (containingDoc === undefined) {
             return;
           }
@@ -72,12 +73,12 @@ class CreateElementExtension extends Rule {
 
   private isExtendingElementCall(expr: babel.Expression): boolean {
     return babel.isCallExpression(expr) &&
-      babel.isMemberExpression(expr.callee) &&
-      babel.isIdentifier(expr.callee.object) &&
-      expr.callee.object.name === 'document' &&
-      babel.isIdentifier(expr.callee.property) &&
-      expr.callee.property.name === 'createElement' &&
-      expr.arguments.length >= 2;
+        babel.isMemberExpression(expr.callee) &&
+        babel.isIdentifier(expr.callee.object) &&
+        expr.callee.object.name === 'document' &&
+        babel.isIdentifier(expr.callee.property) &&
+        expr.callee.property.name === 'createElement' &&
+        expr.arguments.length >= 2;
   }
 }